backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.47
pydantic==1.10.2
asyncpg==0.27.0
databases[postgresql]==0.5.5
psycopg2-binary==2.9.5
alembic==1.11.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
